+++ Wireframe-Admob-All-IN-ONE/app/src/main/java/com/admob/android/ads/withwireframe/bestpractices/model/LayoutItem.kt
@@ -1,0 +1,6 @@
+package com.admob.android.ads.withwireframe.bestpractices.model
+
+data class LayoutItem(
+    val title: String ,
+    val description: String
+)