+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## `04 Jun 2025`
+
+- Wireframe App added for all Admob setups
+- layout improvements
+
 ## `03 Jun 2025`
 
 - Added Admob Banner Size 320x50, 320x100, 468x60, 728x90.